+++ lib/instrumentation/mongodb/common.js
@@ -49,13 +49,14 @@ common.instrumentCollection = function instrumentCollection(shim, Collection) {
  *
  * @param {Shim} shim
  * @param {BulkOperationModule} bulk operation module, typically from mongodb/lib/bulk/common
+ * @param BulkOperationModule
  */
 common.instrumentBulkOperation = function instrumentBulkOperation(shim, BulkOperationModule) {
-  const BulkOperationBase = BulkOperationModule && BulkOperationModule.BulkOperationBase;
+  const BulkOperationBase = BulkOperationModule && BulkOperationModule.BulkOperationBase
 
   if (BulkOperationBase && BulkOperationBase.prototype) {
-    const proto = BulkOperationBase.prototype;
-    shim.recordBatchQuery(proto, 'execute', common.makeBulkDescFunc(shim, 'execute'));
+    const proto = BulkOperationBase.prototype
+    shim.recordBatchQuery(proto, 'execute', common.makeBulkDescFunc(shim, 'execute'))
   }
 }
 
@@ -101,12 +102,12 @@ common.makeQueryDescFunc = function makeQueryDescFunc(shim, methodName) {
 }
 
 /**
- * Sets up the desc for all instrumented query methods
+ * Sets up the desc for all instrumented bulk operations
  *
  * @param {Shim} shim
  * @param {string} methodName
  */
-common.makeBulkDescFunc = function makeBulkDescFunc(shim, methodName) {
+common.makeBulkDescFunc = function makeBulkDescFunc(shim) {
   return function bulkDescFunc() {
     const parameters = getInstanceAttributeParameters(shim, this)
     return {
@@ -116,7 +117,7 @@ common.makeBulkDescFunc = function makeBulkDescFunc(shim, methodName) {
       callback: shim.LAST,
       opaque: true
     }
-  };
+  }
 }
 
 /**
@@ -198,7 +199,8 @@ function setHostPort(shim, connStr, db, client) {
 function getInstanceAttributeParameters(shim, obj) {
   if (obj.s && obj.s.topology) {
     shim.logger.trace('Adding datastore instance attributes from obj.s.db + obj.s.topology')
-    const databaseName = obj.s.db && obj.s.db.databaseName || obj.s.namespace && obj.s.namespace.db || null
+    const databaseName =
+      (obj.s.db && obj.s.db.databaseName) || (obj.s.namespace && obj.s.namespace.db) || null
     const topology = obj.s.topology
     if (topology.s && topology.s.options) {
       return doCapture(topology.s.options, databaseName)
